requirements.txt
# Core Data Processing
numpy>=1.24.0
pandas>=1.5.0
scipy>=1.11.0
pandas-datareader>=0.10.0

# Financial Data
yfinance>=0.2.28
ta>=0.10.2
fredapi>=0.5.0
alpaca-py>=0.7.0

# Machine Learning
scikit-learn>=1.3.0
tensorflow>=2.13.0
xgboost>=1.7.0
lightgbm>=4.1.0
catboost>=1.2.0
scikit-learn-extra>=0.2.0

# Natural Language Processing
nltk>=3.8.1
transformers>=4.36.0
praw>=7.7.0  # Python Reddit API Wrapper

# Web & API
requests>=2.31.0
websocket-client>=1.6.0
websockets>=12.0
fastapi>=0.109.0
uvicorn>=0.27.0
python-multipart>=0.0.6
pydantic>=2.5.0
httpx>=0.26.0
aiohttp>=3.9.1
newsapi-python>=0.2.7
tweepy>=4.13.0

# Visualization
matplotlib>=3.7.0
plotly>=5.18.0
plotly-express>=0.4.0
seaborn>=0.12.0
altair>=5.1.0
bokeh>=3.3.0

# Web App
streamlit>=1.29.0
beautifulsoup4>=4.12.0

# Optimization
cvxpy>=1.3.2

# Statistics & Econometrics
statsmodels>=0.14.0

# Database
redis>=5.0.0
psycopg2-binary>=2.9.9
sqlalchemy>=2.0.25
alembic>=1.12.1
aiosqlite>=0.19.0

# Authentication & Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Development & Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.23.0
black>=23.11.0
flake8>=6.1.0
jupyter>=1.0.0

# Deployment & Monitoring
gunicorn>=21.2.0
sentry-sdk>=1.41.0
prometheus-client>=0.19.0

# Environment
python-dotenv>=1.0.0
